Rewrite how schemeless.relative_to() works
Previously there was specialization code in the file scheme
that tried to work with file and scheme-less URIs. This code
worked except it failed to understand that we need S3 URIs
to be the parent for a scheme-less child.

Change the logic so that file is now a completely standard
implementation using the base class, and schemeless converts
the parameters to handle the different scenarios before calling
the base implementation.