Improve type annotations for findFileResources.
The return type is controlled directly by the value of the 'grouped'
argument, so with typing.Literal and typing.overload we can save
downstream code from constantly having to assert that the return type
is the one it asked for.

In addition, obs.base.RawIngestTask has been assuming it could pass an
re.Pattern in as file_filter, and it's right, because re.compile on an
already-compiled pattern works fine.